requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
babel==2.7.0              # via django-phonenumber-field
backcall==0.1.0           # via ipython
cairocffi==1.1.0          # via cairosvg, weasyprint
cairosvg==2.4.2           # via weasyprint
certifi==2019.9.11        # via requests
cffi==1.13.2              # via cairocffi, weasyprint
chardet==3.0.4            # via requests
cmsplugin-filer==1.1.3
<<<<<<< HEAD
Django==1.8.16
django-appconf==1.0.2
django-authority==0.11 #serve a django-page-cms ma non è incluso nelle sue dipendenze
django-classy-tags==0.8.0
django-cms==3.4.4
djangocms-admin-style==1.2.6
djangocms-grid==1.3
djangocms-text-ckeditor==3.3.0
django-crontab==0.4.2
django-filebrowser-no-grappelli==3.6.2
django-filer==1.2.5
django-markitup==2.2.2
django-mptt==0.8.6
django-pingback==0.3
django-polymorphic==0.8.1
django-rosetta==0.7.13
django-sekizai==0.10.0
django-tagging==0.3.6
django-taggit==0.14.0
django-xmlrpc==0.1.5
easy-thumbnails==2.3
ecdsa==0.13
-e git+https://github.com/dcramer/django-paypal.git#egg=django_paypal-master
gunicorn==19.5.0
html2text==2015.2.18
html5lib==0.95
httplib2==0.9
lxml==3.4.2
markdown2==2.3.5
oauthlib==0.7.2
paramiko==1.17.6
Pillow==2.6.1
polib==1.0.4
Pygments==2.0.2
PyJWT==0.4.1
python-memcached==1.48
python-oembed==0.2.1
python-openid==2.2.5
python-social-auth==0.2.9
python-twitter==2.2
pytz
raven==5.1.1
requests==2.5.3
responses==0.8.1
requests-oauthlib==0.4.2
simplejson==3.6.5
six==1.9.0
stripe==1.19.1
suds==0.4
Unidecode==0.04.17
django-recaptcha==1.0.5
django-contrib-comments==1.7.1
django_extensions==1.6.7
openpyxl==2.4.8
ipython==5.4.1
decorator==4.1.2
=======
cssselect2==0.2.2         # via cairosvg, weasyprint
dataclasses==0.7
decorator==4.4.1
defusedxml==0.6.0         # via cairosvg, python3-openid, social-auth-core
dj-database-url==0.5.0
django-appconf==1.0.3     # via cmsplugin-filer
django-authority==0.13.2
django-classy-tags==0.9.0
django-cms==3.6.0
django-crispy-forms==1.8.0
django-crontab==0.7.1
django-extensions==2.2.5
django-filebrowser-no-grappelli==3.7.8
django-filer==1.5.0
django-formtools==2.1     # via django-cms
django-js-asset==1.2.2    # via django-mptt
django-markitup==3.0.0
>>>>>>> fbe11d2250baeca477a299ff135a1827ec1b9880
django-model-utils==3.1.1
django-mptt==0.10.0       # via django-filer
django-phonenumber-field==3.0.1
django-polymorphic==2.0.3  # via django-filer
django-rosetta==0.9.3
django-sekizai==1.0.0     # via cmsplugin-filer, django-cms
django-tagging==0.4.6
django-taggit-labels==0.6.0
django-taggit==0.22.1
django-treebeard==4.3     # via django-cms
django-xmlrpc==0.1.8
django==1.11.26
djangocms-admin-style==1.4.0
djangocms-attributes-field==1.1.0  # via cmsplugin-filer
djangocms-grid==1.3.0
djangocms-text-ckeditor==3.8.0
easy-thumbnails==2.6      # via cmsplugin-filer, django-filer
ecdsa==0.14.1             # via python-jose
et-xmlfile==1.0.1         # via openpyxl
future==0.18.2            # via pyjwkest, python-jose
gunicorn==19.9.0
html5lib==1.0.1           # via djangocms-text-ckeditor, weasyprint
idna==2.8                 # via requests
ipython-genutils==0.2.0   # via traitlets
ipython==7.9.0
jdcal==1.4.1              # via openpyxl
jedi==0.15.1              # via ipython
jinja2==2.10.3
lxml==4.4.1
markdown2==2.3.8
markupsafe==1.1.1         # via jinja2
oauthlib==3.1.0           # via requests-oauthlib, social-auth-core
openpyxl==2.4.8
parso==0.5.1              # via jedi
pexpect==4.7.0            # via ipython
phonenumbers==8.10.23
pickleshare==0.7.5        # via ipython
pillow==6.2.1             # via cairosvg, djangocms-text-ckeditor, easy-thumbnails
polib==1.1.0              # via django-rosetta
prompt-toolkit==2.0.10    # via ipython
ptyprocess==0.6.0         # via pexpect
pyasn1==0.4.8             # via rsa
pycparser==2.19           # via cffi
pycryptodomex==3.9.4      # via pyjwkest
pygments==2.4.2           # via ipython
pyjwkest==1.4.0
pyjwt==1.7.1              # via social-auth-core
pyphen==0.9.5             # via weasyprint
python-decouple==3.3
python-jose==3.0.1        # via social-auth-core
python-oembed==0.2.4
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.3              # via babel, django
requests-oauthlib==1.3.0
requests==2.22.0
rsa==4.0                  # via python-jose
shortuuid==0.5.0
six==1.13.0               # via django-appconf, django-extensions, django-rosetta, ecdsa, html5lib, prompt-toolkit, pyjwkest, python-jose, social-auth-app-django, social-auth-core, traitlets
social-auth-app-django==3.1.0
social-auth-core[openidconnect]==3.2.0
stripe==2.40.0
tinycss2==1.0.2           # via cairosvg, cssselect2, weasyprint
traitlets==4.3.3          # via ipython
unicodecsv==0.14.1
<<<<<<< HEAD
cairocffi==0.8.1
cffi==1.11.5
CairoSVG==1.0.22
=======
unidecode==1.0.23         # via django-filer
urllib3==1.25.7           # via requests
wcwidth==0.1.7            # via prompt-toolkit
weasyprint==50
webencodings==0.5.1       # via html5lib, tinycss2

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.6.0        # via cairocffi, ipython, tinycss2, weasyprint
>>>>>>> fbe11d2250baeca477a299ff135a1827ec1b9880
